Extensive changes to the Coords class used to track specimen/solvent/fft padding in the simulation. Added methods to handle allocation and resizing. Re-structured the set-up of the solvent molecules in memory in a way to force the OMP scheduling to reduce the number of atomic update clashes. Initial efforts show a factor of 2 speed up on calc-waters, which is about a 20-25% reduction in overall run-time. I think a further improvement can be made by padding each thread chunk to ensure the scheduling doesn't diverge across the spatial boundaries that make the approach fast.

+++ src/core/pdb.h
@@ -91,8 +91,8 @@ class PDB {
 		float average_bFactor;
 		float pixel_size;
 		float vol_angX, vol_angY, vol_angZ;
-		long vol_nX, vol_nY, vol_nZ;
-		long vol_oX, vol_oY, vol_oZ;
+		int vol_nX, vol_nY, vol_nZ;
+		int vol_oX, vol_oY, vol_oZ;
 		float cubic_size;
 		float offset_z;
 		float min_z;
